+++ test/vector_test.cpp
@@ -16,14 +16,14 @@ static void test_constructor_span() {
     assert_equal(vec.dimensions(), 3u);
 }
 
-static void test_conversion() {
+static void test_as_vector() {
     auto vec = Vector({1, 2, 3});
-    auto float_vec = static_cast<std::vector<float>>(vec);
+    auto& float_vec = vec.as_vector();
     assert_equal(float_vec == std::vector<float>{1, 2, 3}, true);
 }
 
 void test_vector() {
     test_constructor_vector();
     test_constructor_span();
-    test_conversion();
+    test_as_vector();
 }